Frontend/requirements_streamlit.txt
streamlit>=1.36
pandas>=2.0
numpy>=1.24
pyarrow>=12.0
plotly>=5.18
statsmodels>=0.14
altair
databricks-sql-connector
python-dotenv
